requirements.txt
asgiref==3.7.2
asttokens==2.4.1
boto3==1.28.59
botocore==1.31.59
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
Django==4.2.7
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
executing==2.0.1
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
prompt-toolkit==3.0.40
psutil==5.9.6
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
